renamed register to register-sheep
