+++ src/app/pages/registration/register-sheep/register-sheep-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { RegisterPage } from './register.page';
+import { RegisterPage } from './register-sheep.page';
 
 const routes: Routes = [
 	{
