+++ src/app/pages/registration/register-sheep/register-sheep.page.scss
